Speedup++
the ntop functions are really expensive, so instead of converting those
pesky 32 bits that represent the source IP to a string using ntop we
keep it as an uint32_t type. The blocklist has been replaced with
blacklist that is split in one part for ipv4 and one for ipv6. IPv4 is
indexed by a the IP address expressed as a 32 bit integer and this
proves to be much much faster than doing it as a string.

Seeing close to 10Mpps on my laptop for this compared to 1.5Mpps before
change.