Do progressive block time
Instead of having a fixed block time we start out with an initial value
(default is 10 seconds) and double that until we reach the maximum
value.

From a performance perspective there is not much point increasing this
any further and I think already 600 seconds is quite a long time. Once
we go out of fast block and start measuring packet rates it will only
take the number of packets configured as pps_burst before we go into
fast block again (provided sources still exceeds threshold). Even with a
burst of thousands of packets, this means that re-examining if a source
still violates a threshold is a rather cheap operation and 600 is
therefore a very conservative value.