More speedtesting
So it seems that a lot of time is spent actually creating the C struct
to hold the prefix. This is done by New_Prefix in patricia.c.

This commit is just a test to lift out the creation of the prefix and
the actual lookup.

With creation + lookup I get something like 4Mpps on my laptop. Creating
once and then just doing a lookup for every packets yields almost twice
that, so it's apparent the creation is fairly expensive. Not very
surprising... memcpy is stupid :/

If we want to go the route of using this patricia trie, I think we
should rewrite the whole thing to specifically suite our purpose, like
being able to pass the uint32_t* pointer to the IPv4 address that
already have in memory and not doing memcpy on that....

I'll leavy this code alone for a while as it seems performance without
current version of patricia is actually better :P