Fix blacklist and mitigation logic
Would fail if a blacklist wasn't defined for the destination IP and this
is certainly not the case since blacklist is more like a dynamic cache
and has no guarantee to be filled in.

Also check if there is a mitigation configured for the destination IP -
if not we just pass the packet. I think this is the only sane thing to
do, like if we are just starting up but the network reroutes packets to
us before we have re-read the configuration we will receive packets to
us without a configured mitigation - should pass.