Periodically reload mitigation configuration
So we re-read the config file periodically and if it has changed we'll
load the new config. Currently no sanity checking so a broken config
file would probably mess things up. Also, re-reading the entire file is
obviously not optimal but it works for now ;)

Also switched the CLI arguments around. I'm thinking we want --config
for the "main" configuration of snabbddos including things like the
interface setup, VLAN tagging and similar. I want to keep these at two
separate files because one affects the app network in Snabb whereas the
mitigation config is only read by the ddos app.