Add selftest showing performance
Not really a selftest in its truest form as it only measures performance
and outputs this to stdout.

I have to give unique names to the apps in the app network during the
different test runs or Snabb will crash, which I find interesting. I was
under the impression engine.configure() should just reconfigure stuff -
maybe limitation in the app? Will it be sad when it's links goes away
(and a new one shows up)?