Have br-address be ordered-by user
The br-address list must be ordered-by user since we have the "br" leaf under softwire that is supposed to reference into br-address through a zero-based index. Overall this is a tricky solution since such zero-based index isn't really a part of YANG but rather the programming language used to hold the instance data. Thus, when we set the software/br leaf we rely on the order in which it exists, to avoid the list being resorted we have to specify that it is to be ordered-by user.

Part of #636.