Elaborate MTU description and add range and units
I extended the description to clarify that this is IP MTU including the IP header but not accounting for lower layer protocols. I think this is what is actually meant but I will admit I have not run through the code to check.

Added units.. perhaps obvious but doesn't hurt. Last but not least we now have a range on the MTU - always good with some sanity checking. What would happen if someone tries to configure with MTU=3 or something silly small?

Part of #636.